chore: refactored parallel transformer
